// backend/src/app/enums/Token.enum.ts
export enum Token {
    INVALID_TOKEN = 'Invalid token',
    TOKEN_EXPIRED = 'Token expired',
    TOKEN_NOT_PROVIDED = 'Token not provided',
    TOKEN_NOT_GENERATED = 'Token not generated',
    TOKEN_MISSING = 'Token missing',
    TOKEN_MALFORMED = 'Token malformed',
    ERROR_TO_GENERATE_TOKEN = 'Error to generate token',
    INVALID_PASSWORD = 'Invalid password',
    ERROR_TO_COMPARE_PASSWORD = 'Error to compare password',
    USER_NOT_FOUND = 'User not found',
    ERROR_TO_FIND_USER = 'Error to find user',
}